code/requirements.txt
# OpenClaw生物学启发记忆引擎 - 依赖项

# 核心依赖（最小化）
numpy>=1.20.0              # 数值计算，用于向量运算
scikit-learn>=1.0.0        # 机器学习算法（可选，用于高级功能）
pandas>=1.3.0              # 数据分析（可选，用于统计和监控）

# 可选依赖 - 根据配置选择安装
snownlp>=0.12.3            # 中文情感分析库（用于情绪评估模块）
faiss-cpu>=1.7.0              # 向量相似性搜索（用于大规模记忆库）
# sentence-transformers>=2.2.0  # 语义嵌入模型（用于高级语义检索）
# annoy>=1.17.0                 # 近似最近邻搜索（轻量级替代）
# chromadb>=0.4.0               # 向量数据库（用于分布式存储）
# pydantic>=2.0.0               # 数据验证和设置管理（可选）

# 开发依赖
pytest>=7.0.0              # 测试框架
black>=23.0.0              # 代码格式化
flake8>=6.0.0              # 代码检查
mypy>=1.0.0                # 类型检查

# 可选的增强依赖（根据需要添加）
# sentence-transformers>=2.2.0  # 文本嵌入向量生成
# transformers>=4.30.0          # 预训练语言模型
# torch>=2.0.0                  # PyTorch深度学习框架
# openai>=1.0.0                 # OpenAI API访问
# tiktoken>=0.5.0               # OpenAI令牌计数

# 注意：目标是最小化依赖，大多数功能使用标准库实现
# 实际部署时应根据具体功能选择依赖项